+++ CHANGELOG.md
@@ -4,13 +4,17 @@ All notable changes to the "oml-vision" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [0.2.0]
+## [v0.2.0]
 
 ### Fixed
 - Tree View shows VSCode error popup with Tree UI bugs https://github.com/opencaesar/oml-vision/pull/5
 
 ### Added
 - Add export to SVG for diagrams https://github.com/opencaesar/oml-vision/pull/3
+- Add VSCode dev container https://github.com/opencaesar/oml-vision/pull/21
+- Add ability to change node and node text color while automatically changing node color based on node type https://github.com/opencaesar/oml-vision/pull/22
+- Add ability to animate edges https://github.com/opencaesar/oml-vision/pull/23
+- Add ability to highlight/unlight connected edges to selected nodes https://github.com/opencaesar/oml-vision/pull/24
 
 ## [0.1.0]